src/detection_engine/requirements.txt
boto3
jmespath
PyYAML
